pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "coinalyze-eth-usd-warehouse"
version = "0.1.0"
description = "Coinalyze-backed PostgreSQL warehouse for ETH USD multi-granularity market data."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1,<2",
  "psycopg[binary]>=3.2,<4",
]

[project.scripts]
coinalyze-eth-usd-warehouse = "realtime_querydb.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
realtime_querydb = ["sql/*.sql"]
